week_7/Google Agent SDK/pyproject.toml
[project]
name = "a2a_simple"
version = "0.1.0"
description = "Simple agent example that only returns Messages"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "a2a-sdk[http-server]>=0.2.5",
    "uvicorn>=0.34.3",
]

[tool.hatch.build.targets.wheel]
packages = ["."]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
